# Conflicts:
#   .gitignore

+++ backend/auction-service/src/main/java/com/turtlecoin/auctionservice/domain/auction/dto/BidRequestDTO.java
@@ -8,5 +8,5 @@ import lombok.*;
 @Builder
 public class BidRequestDTO {
     private Long userId;
-    private Long bidAmount;
+    private Double bidAmount;
 }
